Move types from model to grouper
