+++ pkg/osvscanner/vulnerability_result.go
@@ -46,7 +46,7 @@ func groupResponseBySource(r *output.Reporter, query osv.BatchedQuery, resp *osv
 
 		pkg.Vulnerabilities = response.Vulns
 
-		pkg.Groups = grouper.Group(models.ConvertToIDAliases(pkg.Vulnerabilities))
+		pkg.Groups = grouper.Group(grouper.ConvertVulnerabilityToIDAliases(pkg.Vulnerabilities))
 		groupedBySource[query.Source] = append(groupedBySource[query.Source], pkg)
 	}
